requirements.txt
asgiref==3.7.2
boto3==1.28.37
botocore==1.31.37
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
jmespath==1.0.1
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
